촬영 요청 API 추가

+++ src/main/java/com/example/fever_server_test/dto/request/VideoStartReqDto.java
@@ -5,6 +5,6 @@ import lombok.Data;
 
 @Data
 public class VideoStartReqDto {
-    int equipmentId;
+    Long equipmentId;
     String loginType;
 }
